Require c99_runtime for builtin-convert-1.c
...which was failing on *-elf.  This test is making sure that we narrow
maths calls, e.g. sinh -> sinhf, but we can only do that if we know that
the target has a full c99 libm or if we have other proof that the function
we want to generate is available.

Tested on aarch64-none-elf (where it's skipped) and aarch64-linux-gnu
(where it isn't).

gcc/testsuite/
	* gcc.dg/torture/builtin-convert-1.c: Require c99_runtime.


git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@229548 138bc75d-0d04-0410-961f-82ee72b054a4