rs6000: Optimise SImode cstore on 64-bit
On 64-bit we can do comparisons of 32-bit values by extending those
values to 64-bit, subtracting them, and then getting the high bit of
the result.  For registers this is always cheaper than using the carry
bit sequence; and if the comparison involves a constant, this is cheaper
than the sequence we previously generated in half of the cases (and the
same cost in the other cases).

After this, the only sequence left that is using the mfcr insn is the
one doing signed comparison of Pmode registers.


git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@231165 138bc75d-0d04-0410-961f-82ee72b054a4